requirements.txt
streamlit==1.44.1
kafka-python==2.1.2
filelock==3.18.0
